Moved code from Program.cs to DeploymentFactory.cs
Moved the piece of code that is used to prompt the user for a new Deployment.
